# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kolibri AI OS V2.1",
    "version": "2.1.0"
  },
  "paths": {
...
    "/v1/api-keys": {
      "post": {
        "tags": [
          "api-keys"
        ],
        "summary": "Create Api Key",
        "operationId": "create_api_key_v1_api_keys_post",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/APIKeyCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      },
      "get": {
        "tags": [
          "api-keys"
        ],
        "summary": "List Api Keys",
        "operationId": "list_api_keys_v1_api_keys_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
    "/v1/api-keys/{key_id}": {
      "delete": {
        "tags": [
          "api-keys"
        ],
        "summary": "Revoke Api Key",
        "operationId": "revoke_api_key_v1_api_keys__key_id__delete",
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
...
    "/v1/artifacts": {
      "get": {
        "tags": [
          "artifacts"
        ],
        "summary": "List Artifacts",
        "operationId": "list_artifacts_v1_artifacts_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "name": "estimate_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Estimate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
    "/v1/artifacts/{artifact_id}": {
      "get": {
        "tags": [
          "artifacts"
        ],
        "summary": "Get Artifact",
        "operationId": "get_artifact_v1_artifacts__artifact_id__get",
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
    "/v1/artifacts/{artifact_id}/content": {
      "get": {
        "tags": [
          "artifacts"
        ],
        "summary": "Artifact Content",
        "operationId": "artifact_content_v1_artifacts__artifact_id__content_get",
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIKeyCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": "developer"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "APIKeyCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "sk-kolibri API key or signed Kolibri session",
        "description": "OpenAI-compatible bearer authentication."
      },
      "KolibriSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "kolibri_session",
        "description": "Secure HttpOnly browser session cookie."
      },
      "KolibriSessionHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Kolibri-Session"
      },
      "KolibriNodeToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Kolibri-Node-Token"
      }
    }
  },
...
}